tasa_v2/back/template_microservices/requirements/base.txt
SQLAlchemy==1.3.15
pg8000==1.15.2
Flask-Script==2.0.6
flask_cors==3.0.8
Flask-SQLAlchemy==2.4.1
jsonschema==3.2.0 
py-ms[all]==2.4.1
pysftp==0.2.9
pyjwt==1.7.1